Remove PublicKeyImportService and adapt imports
